Move lower level methods to the createMediaTrait for possible reuse elsewhere.
