+++ docroot/modules/custom/sitenow_migrate/src/Plugin/migrate/CreateMediaTrait.php
@@ -2,6 +2,9 @@
 
 namespace Drupal\sitenow_migrate\Plugin\migrate;
 
+use Drupal\migrate\MigrateException;
+use Drupal\migrate\Row;
+
 /**
  * A trait for handling on-demand media creation.
  */
@@ -118,6 +121,165 @@ trait CreateMediaTrait {
       }
     }
     return FALSE;
+  }
+
+
+  /**
+   * Check if the file exists using the filename.
+   *
+   * @param string $filename
+   *   The filename.
+   *
+   * @return mixed
+   *   Returns the file ID or FALSE if it doesn't exist.
+   */
+  public function getNewFileId($filename) {
+    // Get a connection for the destination database
+    // and retrieve the associated fid.
+    return \Drupal::database()->select('file_managed', 'f')
+      ->fields('f', ['fid'])
+      ->condition('f.filename', $filename)
+      ->execute()
+      ->fetchField();
+  }
+
+  /**
+   * Return the path we'll be writing to.
+   */
+  protected function getDestinationFilePath($filename) {
+    return 'public://' . date('Y-m') . '/' . $filename;
+  }
+
+  /**
+   * Get an existing file entity for a given URI.
+   *
+   * @param string $destination
+   *   The destination URI.
+   * @param bool $set_permanent
+   *   Whether or not to set the file as permanent if it is currently temporary.
+   *
+   * @return \Drupal\file\FileInterface
+   *   The matching file entity or NULL if no entity exists.
+   */
+  protected function getExistingFileEntity($destination, $set_permanent = TRUE) {
+    if ($files = \Drupal::entityTypeManager()->getStorage('file')->loadByProperties(['uri' => $destination])) {
+      // Grab the first file entity with a matching uri.
+      // @todo Any logic for preference when there are multiple?
+      $file = reset($files);
+
+      // Set to permanent if the file in the database is set to temporary.
+      // This means that the file was probably set to be removed during
+      // garbage collection, which we don't want to happen anymore since we're
+      // using it.
+      if ($set_permanent && !$file->isTemporary()) {
+        $file->setPermanent();
+        $file->save();
+      }
+
+      return $file;
+    }
+
+    return NULL;
+  }
+
+  /**
+   * Get the URL of the source public files path with a trailing slash.
+   *
+   * @param \Drupal\migrate\Row $row
+   *   The migration row record.
+   *
+   * @return string
+   *   The URL.
+   *
+   * @throws \Drupal\migrate\MigrateException
+   */
+  protected function getSourcePublicFilesUrl(Row $row): string {
+    if (isset($this->sourceBaseUrl)) {
+      $base_url = rtrim($this->sourceBaseUrl, '/');
+
+      if ($files_dir = $this->getSourceFilePath($row)) {
+        return "{$base_url}/{$files_dir}/";
+      }
+      else {
+        throw new MigrateException('Cannot process media. No public files path variable set.');
+      }
+    }
+    else {
+      throw new MigrateException('Cannot process media. No source base URL set.');
+    }
+  }
+
+  /**
+   * Convenience method to return default scheme for files.
+   */
+  protected function getDefaultScheme() {
+    return \Drupal::config('system.file')
+      ->get('default_scheme');
+  }
+
+  /**
+   * Get the source file path.
+   */
+  protected function getSourceFilePath(Row $row) {
+    return $row->getSourceProperty('constants')['source_file_path'] ?? '';
+  }
+
+  /**
+   * Get an existing media entity ID based on a filename.
+   */
+  public function getMidByFilename($filename, $type = 'image') {
+    // Array of Medias witch contains your file.
+    // Load file by filename
+    // array.
+    $files = $this->entityTypeManager
+      ->getStorage('file')
+      ->getQuery()
+      ->condition('filename', $filename)
+      ->execute();
+
+    if (empty($files)) {
+      return FALSE;
+    }
+
+    // Get First file (make a loop if you get many files)
+    $fileId = array_shift($files);
+
+    // Array of Medias which contains your file.
+    $medias = $this->entityTypeManager
+      ->getStorage('media')
+      ->loadByProperties(['field_media_image' => $fileId]);
+    if (empty($medias)) {
+      return FALSE;
+    }
+    // Again, let's grab the first.
+    $media = array_shift($medias);
+    return $media->id();
+  }
+
+  /**
+   * Check if a source exists.
+   */
+  protected function sourceExists($path) {
+    if ($this->isLocalUri($path)) {
+      return is_file($path);
+    }
+    else {
+      try {
+        $method = !empty($this->configuration['source_check_method']) ? $this->configuration['source_check_method'] : 'HEAD';
+        $options = !empty($this->configuration['guzzle_options']) ? $this->configuration['guzzle_options'] : [];
+        \Drupal::httpClient()->request($method, $path, $options);
+        return TRUE;
+      }
+      catch (ServerException $e) {
+        return FALSE;
+      }
+      catch (ClientException $e) {
+        return FALSE;
+      }
+      catch (ConnectException $e) {
+        return FALSE;
+      }
+    }
   }
 
 }
